Update 20150606
I could have sent this out a day or so earlier, but I was in no real
hurry . . .
Everything should be fully functional now, but there are some things
that we should handle with css/js:
1.  On the delete, the instructions called for a javascript validation
to make sure that the user wanted to delete.  Right now these are links,
not sure if you want to change them to buttons, and attach an event (I
know it can be done on a link as well).  This will need to be applied to
both landing.jsp and detail.jsp
2.  As stated earlier, looking for some classes to hide links/divs etc.
when a user is not logged in.  I did it the long way on the header, but
in other places like landing and edit, I just changed the class.  We can
do it through java/etc, but it is much cleaner if we use java to set the
class, and have CSS hide it.
3.  I disabled the username on the profile update . . .
4. Posts creation/edits need the comment area resized.

That is all that I can think of; see you Tuesday